Create button client component
As it contains client side functionality - onClick
