+++ examples/react18-server-client-library/packages/my-lib/src/client/button/index.ts
@@ -1,0 +1,3 @@
+"use client";
+
+export * from "./button";
