--- examples/react18-server-client-library/packages/my-lib/src/code.tsx
@@ -1,9 +1,0 @@
-export function Code({
-  children,
-  className,
-}: {
-  children: React.ReactNode;
-  className?: string;
-}): JSX.Element {
-  return <code className={className}>{children}</code>;
-}
